Datatypes: use diamond operator, remove redundant initialization values, convert comments

+++ src/org/infinity/datatype/Song2daBitmap.java
@@ -88,7 +88,7 @@ public class Song2daBitmap extends ResourceBitmap
         if (defValue.equalsIgnoreCase(ref)) {
           ref = "None";
         }
-        SongList.add(new RefEntry((long)i, ref, name, searchDirs));
+        SongList.add(new RefEntry(i, ref, name, searchDirs));
       }
     }
     return SongList;
